Fix header standardization
Fix header standardization method and refactor header and query parameters to separate classes
